merged-app/merged-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pydantic
google-auth
requests
python-dotenv
